requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.7.4
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
Django==5.0.6
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jsonpointer==3.0.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cudnn-cu11==9.2.1.18
orjson==3.10.6
packaging==24.1
parso==0.8.4
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
PyAutoGUI==0.9.54
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
PyYAML==6.0.1
regex==2024.7.24
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.13.2
tenacity==8.5.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
